refactor: @components/ 뒤로 붙어있는 import 경로 제거 후 @components로 수정
