+++ client/src/components/ConfirmDialog/ConfirmDialog.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import { useDialog } from '@hooks/useDialog';
-import { Dialog } from '@components/_common';
-import { Text } from '@components/_common/Text';
+import { Dialog, Text, Button } from '@components';
 import Alert from '@assets/images/alert.svg';
 import Check from '@assets/images/check.svg';
-import { Button } from '@components/_common/Button';
 import { confirmDialogHeaderContainer } from './ConfirmDialog.style';
 import type { DialogProps } from './types';
 import type { ButtonSize } from '@components/_common/Button';
