+++ client/src/components/_common/Card/DescriptionText.tsx
@@ -1,4 +1,4 @@
-import { Text } from '@components/_common/Text';
+import { Text } from '@components';
 import React from 'react';
 
 interface DescriptionTextProps {
