+++ client/src/components/_common/Dialog/Dialog.tsx
@@ -8,7 +8,7 @@ import {
     actionContainer,
 } from './Dialog.style';
 import XIcon from '@assets/images/xIcon.svg';
-import { Button } from '@components/_common/Button';
+import { Button } from '@components';
 import type {
     DialogHeaderProps,
     DialogContentProps,
